requirements.txt
Flask
opencv-python
numpy
Pillow
Werkzeug
python-dotenv